pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gigiac-starter-bot"
version = "1.0.0"
description = "Reference bot implementation for the Gigiac AI task marketplace"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.scripts]
gigiac-bot = "gigiac_bot.__main__:main"
